made clean task work with reps that have no raw path

+++ lib/nanoc3/tasks/clean.rb
@@ -14,7 +14,7 @@ module Nanoc3::Tasks
 
       # Delete all compiled item reps
       filenames.each do |filename|
-        FileUtils.rm_f filename
+        FileUtils.rm_f filename unless filename.nil?
       end
     end
